requirements.txt
fastapi
sqlalchemy
alembic
psycopg2-binary
python-dotenv
fastapi-admin
uvicorn
pydantic